refactor: Mirror the test package tree onto the new layout
Move each test class into the package of the class it exercises:
- controller tests + AbstractContentItemControllerTests -> contentitem.rest
- ContentItemMapperTest -> contentitem.model
- the generic ES (de)serializer and discovery tests -> top-level
  elasticsearch (their subjects moved there)

Service, content-aware ES (indexer/search), transformer, and the shared
testutils packages stay put. Imports updated accordingly; no behaviour
change.
