+++ src/test/java/com/contented/contented/contentitem/model/ContentItemMapperTest.java
@@ -1,4 +1,4 @@
-package com.contented.contented.contentitem;
+package com.contented.contented.contentitem.model;
 
 import com.contented.contented.common.UuidV7;
 import com.contented.contented.contentitem.model.ContentItemDTO;
