+++ src/test/java/com/contented/contented/contentitem/rest/ContentItemControllerSearchIndexTests.java
@@ -1,4 +1,4 @@
-package com.contented.contented.contentitem;
+package com.contented.contented.contentitem.rest;
 
 import com.contented.contented.elasticsearch.ElasticSearchIndexCreator;
 import com.contented.contented.contentitem.model.ContentItemResponseDTO;
